refactor(web): extract mode combobox, hide duplicate model description
